refactor: move some exports around
