+++ packages/captp/lib/captp.js
@@ -3,7 +3,8 @@
 // This logic was mostly lifted from @agoric/swingset-vat liveSlots.js
 // Defects in it are mfig's fault.
 import { Remotable, makeMarshal, QCLASS } from '@agoric/marshal';
-import { E, HandledPromise } from '@agoric/eventual-send/shim';
+import { HandledPromise } from '@agoric/eventual-send/shim';
+import { E } from '@agoric/eventual-send';
 import { isPromise } from '@agoric/promise-kit';
 
 export { E, HandledPromise };
