+++ packages/eventual-send/shim.js
@@ -1,1 +1,2 @@
-export * from './src/index';
+// eslint-disable-next-line import/prefer-default-export
+export { HandledPromise } from './src/index';
