+++ packages/eventual-send/src/E.js
@@ -1,5 +1,9 @@
 /* global harden */
 
+// FIXME: Remove this and just use harden once the E proxy no longer
+// is created by ../shim.js.
+const FIXMEharden = x => x;
+
 // eslint-disable-next-line spaced-comment
 /// <reference path="index.d.ts" />
 
@@ -71,5 +75,5 @@ export default function makeE(HandledPromise) {
   E.when = (x, onfulfilled = undefined, onrejected = undefined) =>
     HandledPromise.resolve(x).then(onfulfilled, onrejected);
 
-  return Object.freeze(E);
+  return FIXMEharden(E);
 }
