+++ packages/eventual-send/src/index.js
@@ -23,11 +23,10 @@ const hp =
   typeof HandledPromise === 'undefined'
     ? // eslint-disable-next-line no-use-before-define
       makeHandledPromise(Promise)
-    : harden(HandledPromise);
+    : HandledPromise;
 
 // Provide our exports.
 export { hp as HandledPromise };
-export const E = makeE(hp);
 
 // the following method (makeHandledPromise) is part
 // of the shim, and will not be exported by the module once the feature
