+++ packages/eventual-send/src/no-shim.js
@@ -1,1 +1,5 @@
-export * from './index';
+/* global HandledPromise */
+import makeE from './E';
+
+// eslint-disable-next-line import/prefer-default-export
+export const E = makeE(HandledPromise);
