+++ packages/eventual-send/test/get-hp.js
@@ -1,5 +1,5 @@
 import makeE from '../src/E';
-import makeHandledPromise from '../src/handled-promise';
+import { makeHandledPromise } from '../src/index';
 
 export const HandledPromise = makeHandledPromise(Promise);
 export const E = makeE(HandledPromise);
