+++ packages/eventual-send/test/test-e.js
@@ -1,6 +1,6 @@
 import '@agoric/install-ses';
 import test from 'tape-promise/tape';
-import { E, HandledPromise } from '../src/index';
+import { E, HandledPromise } from './get-hp';
 
 test('E reexports', async t => {
   try {
